+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Core SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Core SDK version within that major version.
 
+### Version 0.2.8.1.4 *(2023-11-29)*
+- This version of the adapter is compatible with Chartboost Core 0.4.0
+- Added the ability to set a Template ID to Mediation Partner ID map to facilitate per-partner consent.
+
 ### 0.2.8.1.3
-- This version of the adapter is compatible with Chartboost Core 0.3.0+.
+- This version of the adapter is compatible with Chartboost Core 0.3.0 only.
 
 ### 0.2.8.1.2
 - Fix a bug around callbacks not firing when consent changes.